# openapi-ci-api.yaml
openapi: 3.0.3
info:
  title: CCExtractor CI System API
  version: 1.2.0
  description: |
    Security-hardened JSON-only REST API for the CCExtractor CI/sample platform.
    Designed for AI agents and CI automation. Enforces scoped Bearer token auth,
    strict input validation, rate limiting on all routes, and safe defaults
    throughout. No browser sessions, no HTML, no implicit permissions.

    **Authentication:** All endpoints require bearer token authentication unless
    explicitly marked with `security: []` (only /system/health and POST /auth/tokens).

    **Rate-limit headers:** Every response includes `X-RateLimit-Limit`,
    `X-RateLimit-Remaining`, and `X-RateLimit-Reset` headers. These are modeled
    explicitly on the 429 response for brevity; they are present on all responses
    regardless of status code.

  contact:
    name: CCExtractor Development
    url: https://github.com/CCExtractor/sample-platform
  license:
    name: GPL-3.0-only
    url: https://www.gnu.org/licenses/gpl-3.0.html
servers:
  - url: http://localhost:5000/api/v1
    description: Local development server
  - url: https://sampleplatform.ccextractor.org/api/v1
    description: Production

#
# Global security: all endpoints require auth
# unless explicitly overridden with security: []
#
security:
  - bearerAuth: []
...
paths:
...
  /runs/{run_id}/samples/{sample_id}/regression-tests/{regression_id}/outputs/{output_id}/expected:
    get:
      tags: [Results]
      summary: Get expected output for a regression test result
      operationId: getExpectedOutput
      description: >
        Expected output is a file reference stored under TestResults using the
        regression output extension. Resolved from GCS or local
        SAMPLE_REPOSITORY at request time. storage_status reflects which
        backends have the file. Do not assume local and GCS are always in sync.
      security:
        - bearerAuth: []
      x-required-scope: results:read
      parameters:
        - $ref: "#/components/parameters/RunId"
        - $ref: "#/components/parameters/SampleId"
        - $ref: "#/components/parameters/RegressionId"
        - $ref: "#/components/parameters/OutputId"
        - $ref: "#/components/parameters/Format"
      responses:
        "200":
          description: Expected output file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OutputFile"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        default:
          $ref: "#/components/responses/Error"
  /runs/{run_id}/samples/{sample_id}/regression-tests/{regression_id}/outputs/{output_id}/actual:
    get:
      tags: [Results]
      summary: Get actual output generated by a regression test in a run
      operationId: getActualOutput
      description: >
        IMPORTANT: TestResultFile.got = null means the actual output MATCHED
        expected, not that actual output is missing. This is a semantic trap
        in the data model. Missing output is represented by a dummy row
        (-1,-1,-1,'','error') which the API translates to status=missing_output
        and returns 404. A 200 response always contains a real output file.
      security:
        - bearerAuth: []
      x-required-scope: results:read
      parameters:
        - $ref: "#/components/parameters/RunId"
        - $ref: "#/components/parameters/SampleId"
        - $ref: "#/components/parameters/RegressionId"
        - $ref: "#/components/parameters/OutputId"
        - $ref: "#/components/parameters/Format"
      responses:
        "200":
          description: Actual output file (output exists and differs from expected)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OutputFile"
        "303":
          description: Output matched expected. Redirected to /expected.
          headers:
            Location:
              schema:
                type: string
                format: uri
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        default:
          $ref: "#/components/responses/Error"
  /runs/{run_id}/samples/{sample_id}/regression-tests/{regression_id}/outputs/{output_id}/diff:
    get:
      tags: [Results]
      summary: Get expected-vs-actual diff for a failing regression test result
      operationId: getDiff
      description: >
        The legacy diff route is header-gated (X-Requested-With: XMLHttpRequest),
        not role-gated. The 403 seen on direct browser requests was a
        header-check artifact. This endpoint wraps the XHR logic and returns
        structured JSON — no HTML, no 50-line truncation.
      security:
        - bearerAuth: []
      x-required-scope: results:read
      parameters:
        - $ref: "#/components/parameters/RunId"
        - $ref: "#/components/parameters/SampleId"
        - $ref: "#/components/parameters/RegressionId"
        - $ref: "#/components/parameters/OutputId"
        - name: context_lines
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 3
        - name: format
          in: query
          schema:
            type: string
            enum: [structured, unified]
            default: structured
      responses:
        "200":
          description: Structured or unified diff
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Diff"
                  - $ref: "#/components/schemas/UnifiedDiff"
                discriminator:
                  propertyName: format
                  mapping:
                    structured: "#/components/schemas/Diff"
                    unified: "#/components/schemas/UnifiedDiff"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        default:
          $ref: "#/components/responses/Error"
  /runs/{run_id}/samples/{sample_id}/baseline-approval:
    post:
      tags: [Results]
      summary: Approve actual output as new expected baseline
      operationId: approveBaseline
      description: >
        Requires baselines:write scope and admin role.
        This is a destructive write — the approved output becomes the new
        expected baseline for the regression test.
      security:
        - bearerAuth: []
      x-required-scope: baselines:write
      x-required-roles: [admin, contributor]
      parameters:
        - $ref: "#/components/parameters/RunId"
        - $ref: "#/components/parameters/SampleId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BaselineApprovalRequest"
      responses:
        "200":
          description: Baseline approval applied immediately.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BaselineApproval"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        default:
          $ref: "#/components/responses/Error"

  # ERRORS AND LOGS
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: >
        Opaque server-side API token. Obtain via POST /auth/tokens.
        The CI worker token used by /ci/progress-reporter is a separate
        secret and is NOT valid here. Never use browser session cookies
        for API clients.

  # HEADERS
  headers:
    RateLimitLimit:
      description: Maximum requests allowed in the current window
      schema:
        type: integer
        example: 120
    RateLimitRemaining:
      description: Requests remaining in the current window
      schema:
        type: integer
        example: 117
    RateLimitReset:
      description: Unix timestamp when the rate limit window resets
      schema:
        type: integer
        example: 1748908800

  # PARAMETERS
  parameters:
...
    RunId:
      name: run_id
      in: path
      required: true
      description: Numeric run ID
      schema:
        type: integer
        minimum: 1
    SampleId:
      name: sample_id
      in: path
      required: true
      description: Numeric media sample ID
      schema:
        type: integer
        minimum: 1
...
    RegressionId:
      name: regression_id
      in: path
      required: true
      description: Regression test definition ID
      schema:
        type: integer
        minimum: 1
    OutputId:
      name: output_id
      in: path
      required: true
      description: Output file ID within a regression test definition
      schema:
        type: integer
        minimum: 1
    Format:
      name: format
      in: query
      description: >
        Content encoding for file responses.
        Use text only when the file is known to be UTF-8 compatible.
        Binary or unknown content defaults to base64.
      schema:
        type: string
        enum: [text, base64]
        default: base64

  # RESPONSES
  responses:
    BadRequest:
      description: Request body or query parameters failed schema validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: validation_error
            message: Request failed schema validation.
            details:
              fields:
                commit_sha: Must match pattern ^[a-fA-F0-9]{40}$
                platform: Must be one of [linux, windows]
    Unauthorized:
      description: Missing, expired, or invalid bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: unauthorized
            message: Bearer token is missing, expired, or invalid.
            details: {}
    Forbidden:
      description: Token is valid but lacks the required scope or role
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: forbidden
            message: Token does not have the required scope for this operation.
            details:
              required_scope: runs:write
              token_scopes: [runs:read, results:read]
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: not_found
            message: Run 9317 not found.
            details:
              resource: run
              id: 9317
...
    RateLimited:
      description: Too many requests. Retry after the indicated number of seconds.
      headers:
        Retry-After:
          description: Seconds to wait before retrying
          schema:
            type: integer
            example: 30
        X-RateLimit-Limit:
          $ref: "#/components/headers/RateLimitLimit"
        X-RateLimit-Remaining:
          $ref: "#/components/headers/RateLimitRemaining"
        X-RateLimit-Reset:
          $ref: "#/components/headers/RateLimitReset"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: rate_limited
            message: Rate limit exceeded. Retry after 30 seconds.
            details:
              retry_after: 30
              limit: 120
              window: 60s
    Error:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  # SCHEMAS
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, message, details]
      properties:
        code:
          type: string
          maxLength: 100
          description: Machine-readable error code (snake_case)
          example: not_found
        message:
          type: string
          maxLength: 500
          description: Human-readable error summary
          example: Run 9317 not found.
        details:
          type: object
          additionalProperties: true
          description: >
            Structured context for the error. Always an object, never null.
            Empty object {} when no additional detail is available.
...
    OutputFile:
      type: object
      required: [sample_id, regression_id, output_id, filename, content_type, encoding, content, storage_status]
      properties:
        run_id:
          type: integer
          minimum: 1
          nullable: true
          description: Null for expected output not tied to a specific run.
        sample_id:
          type: integer
          minimum: 1
        regression_id:
          type: integer
          minimum: 1
        output_id:
          type: integer
          minimum: 1
        filename:
          type: string
          maxLength: 255
        content_type:
          type: string
          maxLength: 100
        encoding:
          type: string
          enum: [utf-8, base64]
          description: >
            utf-8 only when file is confirmed text. Default is base64.
        content:
          type: string
          maxLength: 1048576
          description: >
            File content. Base64-encoded unless encoding=utf-8.
            Files exceeding 1MB are truncated. Check truncated=true and use
            download_url for the full file.
        truncated:
          type: boolean
          description: True if content was truncated due to size limits.
        download_url:
          type: string
          format: uri
          nullable: true
          description: URL to download the full file if it was truncated.
        sha256:
          type: string
          pattern: '^[a-fA-F0-9]{64}$'
        storage_status:
          type: string
          enum: [ok, degraded, missing]
          description: >
            ok = file verified in at least one storage backend.
            degraded = file exists but integrity or redundancy check failed.
            missing = file not found in any storage backend.
    Diff:
      type: object
      required: [run_id, sample_id, regression_id, output_id, status]
      properties:
        run_id:
          type: integer
          minimum: 1
        sample_id:
          type: integer
          minimum: 1
        regression_id:
          type: integer
          minimum: 1
        output_id:
          type: integer
          minimum: 1
        status:
          type: string
          enum: [identical, different, missing_expected, missing_actual]
        summary:
          type: object
          required: [added_lines, removed_lines, changed_hunks]
          properties:
            added_lines:
              type: integer
              minimum: 0
            removed_lines:
              type: integer
              minimum: 0
            changed_hunks:
              type: integer
              minimum: 0
        hunks:
          type: array
          maxItems: 500
          items:
            type: object
            required: [expected_start, actual_start, lines]
            additionalProperties: false
            properties:
              expected_start:
                type: integer
                minimum: 0
              actual_start:
                type: integer
                minimum: 0
              lines:
                type: array
                maxItems: 500
                items:
                  type: object
                  required: [kind, text]
                  additionalProperties: false
                  properties:
                    kind:
                      type: string
                      enum: [context, added, removed]
                    expected_line:
                      type: integer
                      minimum: 0
                      nullable: true
                    actual_line:
                      type: integer
                      minimum: 0
                      nullable: true
                    text:
                      type: string
                      maxLength: 1000
    UnifiedDiff:
      type: object
      required: [run_id, sample_id, regression_id, output_id, format, content]
      properties:
        run_id:
          type: integer
        sample_id:
          type: integer
        regression_id:
          type: integer
        output_id:
          type: integer
        format:
          type: string
          enum: [unified]
        content:
          type: string
          description: Raw unified diff text.
          maxLength: 524288
    BaselineApprovalRequest:
      type: object
      required: [regression_id, output_id]
      additionalProperties: false
      properties:
        regression_id:
          type: integer
          minimum: 1
        output_id:
          type: integer
          minimum: 1
        remove_variants:
          type: boolean
          default: false
          description: >
            If true, removes all platform-specific variants (output_id != 1)
            and promotes this output to the global baseline.
    BaselineApproval:
      type: object
      required: [status, run_id, sample_id, regression_id, output_id, requested_by, created_at]
      properties:
        status:
          type: string
          enum: [approved]
        run_id:
          type: integer
          minimum: 1
        sample_id:
          type: integer
          minimum: 1
        regression_id:
          type: integer
          minimum: 1
        output_id:
          type: integer
          minimum: 1
        requested_by:
          type: string
          maxLength: 100
          description: Display name of the user who requested the approval.
        created_at:
          type: string
          format: date-time
...